requirements.txt
ffmpeg-python
webrtcvad-wheels
noisereduce
pydub
soundfile
numpy
librosa
pyrnnoise
